+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [1.2.0] - 2025-08-21
+### Improvements
+- Implemented security middleware (OWASP) with HSTS and CSP configuration ([#24]).
+- Accessibility enhancements (RGAA) ([#25]).
+
+
 ## [1.1.0] - 2025-08-19
 ### Added
 - Support contact form for users to directly reach the support team.